Update package-lock.json
reversal of 4.0.1 with 4.1.0 version of library, also change the library of the remaining functions and redirect the same to the new library and then engage the libraries in the latest bucket. the reversal of 4.0.1 with the 4.1.0 of library will change the package version of the json file so therefore the package must be signed with a new key made inside the microsoft visual studio code, and then resign with android studio.